requirements.txt
chromadb>=0.4.18
openai>=1.3.0
google-generativeai>=0.3.0
click>=8.1.0
rich>=13.7.0
pydantic>=2.5.0
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Build dependencies
pyinstaller>=6.0.0 